Improve design decision for data types and fixed GCC conversion warnings
Add bme280_temp_init flag and automatic call

Update memory usage report

+++ app/fault_manager/fault_manager.c
@@ -50,7 +50,7 @@ static inline void fault_manager_buzz_pullup(void) {
 }
 
 static inline void fault_manager_buzz_pulldown(void) {
-	D14_D19_DATA_REG &= ~BUZZER_OUTPUT;
+	D14_D19_DATA_REG &= ((uint8_t) ~BUZZER_OUTPUT);
 }
 
 static void fault_manager_buzzer(uint16_t duration_s) {
